% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_devolatilized_returns}
\alias{calc_devolatilized_returns}
\title{Calculate devolatilized returns for GARCH(1,1).}
\usage{
calc_devolatilized_returns(rt, eta_df)
}
\arguments{
\item{rt}{returns}

\item{eta}{GARCH parameters}
}
\value{
devolatilized returns
}
\description{
Calculate devolatilized returns for GARCH(1,1).
}
