% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_Dt}
\alias{calc_Dt}
\title{Calculate Dt for GARCH(1,1), a diagonal matrix containing estimated 
volatilities for each series in each time.}
\usage{
calc_Dt(rt, eta_df)
}
\arguments{
\item{rt}{returns}

\item{eta}{GARCH parameters}
}
\value{
DataFrame of volatilities (Dt)
}
\description{
Calculate Dt for GARCH(1,1), a diagonal matrix containing estimated 
volatilities for each series in each time.
}
