% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimateCDCC}
\alias{estimateCDCC}
\title{Estimate GARCH(1,1)-cDCC(1, 1) parameters}
\usage{
estimateCDCC(rt, ini_par = c(0.05, 0.93))
}
\arguments{
\item{rt}{return matrix}
}
\value{
A list containing the estimated parameters, standardized residuals 
and conditional variance for each series
}
\description{
Estimate GARCH(1,1)-cDCC(1, 1) parameters
}
