% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fc}
\alias{fc}
\title{Calculate correction factor}
\usage{
fc(delta, N)
}
\arguments{
\item{delta}{Quantile for Chi-Squared}

\item{N}{Dimension}
}
\value{
Correction factor for the weighted variance estimator
}
\description{
Calculate correction factor
}
