% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geral_calc_portfolio_variance}
\alias{geral_calc_portfolio_variance}
\title{Evaluate cDCC(1, 1) and BIP-cDCC(1, 1) estimators on the estimation of the
weights for the minimum variance portfolio (MVP). The estimated weights are 
compared with the real ones using three performances criteria: 
Frobenius distance, Ratio of variance of the selected MVP and Percentage 
increase of the variance of the selected MVP in relation to the variance of 
the true MVP.}
\usage{
geral_calc_portfolio_variance(
  phi,
  q_phi,
  r_phi,
  rt,
  q_rt,
  r_rt,
  Dt,
  q_Dt,
  r_Dt,
  S,
  q_S,
  r_S
)
}
\arguments{
\item{phi}{true parameter for the cDCC(1,1)}

\item{q_phi}{estimated parameter for the cDCC(1,1)}

\item{r_phi}{estimated parameter for the BIP-cDCC(1,1)}

\item{rt}{returns matrix}

\item{q_rt}{standardized returns by estimated GARCH(1,1) parameters}

\item{r_rt}{standardized returns by estimated BIP-GARCH(1,1) parameters}

\item{Dt}{matrix of conditional variance for each series}

\item{q_Dt}{matrix of estimated conditional variance for each series by 
GARCH(1,1) Quasi-Maximum Likelihood estimator}

\item{r_Dt}{matrix of estimated conditional variance for each series by 
BIP-GARCH(1,1) robust estimator}

\item{S}{unconditional covariance matrix}

\item{q_S}{estimated unconditional covariance matrix for the 
Quasi-Maximum Likelihood estimator}

\item{r_S}{estimated unconditional covariance matrix for the 
robust estimator}

\item{delta}{chi-squared quantile for the definition of the weighting scheme}
}
\value{
MVP performance measures.
}
\description{
Evaluate cDCC(1, 1) and BIP-cDCC(1, 1) estimators on the estimation of the
weights for the minimum variance portfolio (MVP). The estimated weights are 
compared with the real ones using three performances criteria: 
Frobenius distance, Ratio of variance of the selected MVP and Percentage 
increase of the variance of the selected MVP in relation to the variance of 
the true MVP.
}
