% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{high_dimension_estimateCDCC}
\alias{high_dimension_estimateCDCC}
\title{Estimate cDCC(1, 1) parameters}
\usage{
high_dimension_estimateCDCC(rt)
}
\arguments{
\item{rt}{returns matrix}
}
\value{
The estimated parameters for cDCC(1,1)
}
\description{
Estimate cDCC(1, 1) parameters
}
