% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{high_dimension_robust_estimateCDCC}
\alias{high_dimension_robust_estimateCDCC}
\title{Estimate BIP-cDCC(1, 1) parameters}
\usage{
high_dimension_robust_estimateCDCC(rt, delta, k = 30)
}
\arguments{
\item{rt}{returns matrix}

\item{delta}{reweighted unconditional correlation local window size}
}
\value{
The estimated parameters for BIP_GARCH(1,1)-BIP-cDCC(1,1)
}
\description{
Estimate BIP-cDCC(1, 1) parameters
}
