\name{robcdcc-package}
\alias{robcdcc-package}
\alias{robcdcc}
\docType{package}
\title{
\packageTitle{robcdcc}
}
\description{
\packageDescription{robcdcc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{robcdcc}
\packageIndices{robcdcc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{robcdcc}

Maintainer: \packageMaintainer{robcdcc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
