% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robust_calc_Dt}
\alias{robust_calc_Dt}
\title{Calculate Dt for BIP-GARCH(1,1), a diagonal matrix containing estimated 
volatilities for each series in each time.}
\usage{
robust_calc_Dt(rt, eta_df, delta = 0.975)
}
\arguments{
\item{rt}{returns}

\item{eta_df}{BIP-GARCH parameters dataframe}

\item{delta}{BIP-GARCH parameters}
}
\value{
DataFrame of volatilities (Dt) and volatility filter values for 
each time
}
\description{
Calculate Dt for BIP-GARCH(1,1), a diagonal matrix containing estimated 
volatilities for each series in each time.
}
