% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robust_calc_ht}
\alias{robust_calc_ht}
\title{Calculate ht using BIP-GARCH}
\usage{
robust_calc_ht(rt, eta, delta)
}
\arguments{
\item{rt}{returns}

\item{eta}{GARCH parameters}

\item{delta}{Quantile for Chi-Squared}
}
\value{
volatility (ht)
}
\description{
Calculate ht using BIP-GARCH
}
