% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robust_devolatilized_returns}
\alias{robust_devolatilized_returns}
\title{Calculate devolatilized returns using BIP-GARCH especification}
\usage{
robust_devolatilized_returns(rt, eta_df, delta = 0.975)
}
\arguments{
\item{rt}{returns}

\item{eta_df}{GARCH parameters}

\item{delta}{Quantile for Chi-Squared}
}
\value{
devolatilized returns
}
\description{
Calculate devolatilized returns using BIP-GARCH especification
}
