% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robust_estimateCDCC}
\alias{robust_estimateCDCC}
\title{Estimate BIP-cDCC(1, 1) parameters}
\usage{
robust_estimateCDCC(rt, delta = 0.975, k = 30, ini_par = c(0.05, 0.93))
}
\arguments{
\item{rt}{returns matrix}

\item{delta}{chi-squared quantile for the definition of the weighting scheme}

\item{k}{size of the local window for the reweighted mean vector and 
covariance matrix estimators}

\item{ini_par}{initial values for the parameters to be optimized over.}
}
\value{
A list containing the estimated parameters for 
BIP_GARCH(1,1)-BIP-cDCC(1,1), standardized residuals and conditional 
variance for each series
}
\description{
Estimate BIP-cDCC(1, 1) parameters
}
