% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robust_estimateGARCH}
\alias{robust_estimateGARCH}
\title{Estimate GARCH(1,1) parameters using BIP-GARCH specification and 
M-estimation}
\usage{
robust_estimateGARCH(rt, cy, chisq, k = 30)
}
\arguments{
\item{rt}{returns}

\item{cy}{correction factor}

\item{chisq}{Quantile for Chi-Squared}
}
\value{
devolatilized returns
}
\description{
Estimate GARCH(1,1) parameters using BIP-GARCH specification and 
M-estimation
}
