% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCDCC.R
\name{simCDCC}
\alias{simCDCC}
\title{Simulate cDCC(1,1)}
\usage{
simCDCC(phi, eta, S, nobs, ndim, seed)
}
\arguments{
\item{phi}{Parameters for cDCC(1,1)}

\item{eta}{Parameters for GARCH(1,1)}

\item{S}{Intercept matrix}

\item{nobs}{Number of observations}

\item{ndim}{Number of dimensions}

\item{seed}{seed for random sampling}
}
\value{
List with results
}
\description{
Simulate cDCC(1,1)
}
