% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_ht}
\alias{calc_ht}
\title{Calculate ht using GARCH}
\usage{
calc_ht(rt, eta)
}
\arguments{
\item{rt}{returns}

\item{eta}{GARCH parameters}
}
\value{
volatility (ht)
}
\description{
Calculate ht using GARCH
}
